Added Tab to the list view side
