+++ Chess.View/Window/MainWindow.xaml
@@ -342,33 +342,51 @@
                     </Canvas>
                 </Border>
             </Viewbox>
+
             <GridSplitter Grid.Column="1" Width="5" HorizontalAlignment="Stretch" 
                           DragCompleted="GridSplitter_DragCompleted" />
-            <ListView Grid.Column="2" x:Name="myListView" 
+
+            <TabControl Grid.Column="2">
+                <TabItem>
+                    <TabItem.Header>
+                        <StackPanel Orientation="Horizontal">
+                            <Ellipse Width="10" Height="10" Fill="Green"/>
+                            <TextBlock Width="50" Height="30" FontSize="20" >Tab 1</TextBlock>
+                        </StackPanel>
+                    </TabItem.Header>
+                    <StackPanel Orientation="Horizontal">
+                        <ListView Grid.Column="2" x:Name="myListView" 
                         ItemsSource="{Binding Path=Board.ChessMoveSequence.ChessMoves}"
                         Height="Auto" Margin="10" Width="Auto"
                         VerticalAlignment="Top" Background="AliceBlue">
-                <ListView.Resources>
-                    <select:PieceSelectorForMoveSequenceListView x:Key="pieceSelectorForMoveSequenceListView" />
-                </ListView.Resources>
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Orientation="Horizontal">
-                            <TextBlock VerticalAlignment="Center" Text="{Binding MoveNumber}" Margin="10" FontWeight="Bold"/>
-                            <ContentControl Content="{Binding}">
-                                <ContentControl.ContentTemplateSelector>
-                                    <StaticResource ResourceKey="pieceSelectorForMoveSequenceListView" />
-                                </ContentControl.ContentTemplateSelector>
-                            </ContentControl>
-                            <TextBlock VerticalAlignment="Center" Text="{Binding PieceColor}" Margin="10" />
-                            <TextBlock VerticalAlignment="Center" Text="{Binding PieceName}" Margin="10" />
-                            <TextBlock VerticalAlignment="Center" Text="{Binding SourceString}" Margin="10"/>
-                            <TextBlock VerticalAlignment="Center" Text="{Binding TargetString}" Margin="10"/>
-                            <TextBlock VerticalAlignment="Center" Text="{Binding ShortDescription}" Margin="10"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>
+                            <ListView.Resources>
+                                <select:PieceSelectorForMoveSequenceListView x:Key="pieceSelectorForMoveSequenceListView" />
+                            </ListView.Resources>
+                            <ListView.ItemTemplate>
+                                <DataTemplate>
+                                    <StackPanel Orientation="Horizontal">
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding MoveNumber}" Margin="10" FontWeight="Bold"/>
+                                        <ContentControl Content="{Binding}">
+                                            <ContentControl.ContentTemplateSelector>
+                                                <StaticResource ResourceKey="pieceSelectorForMoveSequenceListView" />
+                                            </ContentControl.ContentTemplateSelector>
+                                        </ContentControl>
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding PieceColor}" Margin="10" />
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding PieceName}" Margin="10" />
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding SourceString}" Margin="10"/>
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding TargetString}" Margin="10"/>
+                                        <TextBlock VerticalAlignment="Center" Text="{Binding ShortDescription}" Margin="10"/>
+                                    </StackPanel>
+                                </DataTemplate>
+                            </ListView.ItemTemplate>
+                        </ListView>
+                    </StackPanel>
+                </TabItem>
+                <TabItem Header="Tab 2">
+                    <!--Bind TextBlock.Text to the TextBox on the first TabItem.-->
+                    <TextBlock Text="Hee"/>
+                </TabItem>
+            </TabControl>
         </Grid>
     </DockPanel>
 </mah:MetroWindow>
